+++ executors/controls.py
@@ -21,7 +21,7 @@ from modules.database import database
 from modules.exceptions import StopSignal
 from modules.logger.custom_logger import logger
 from modules.models import models
-from modules.utils import shared, support, util
+from modules.utils import shared, support
 
 db = database.Database(database=models.fileio.base_db)
 ram = support.size_converter(byte_size=models.settings.ram).replace('.0', '')
@@ -90,7 +90,7 @@ def exit_process() -> NoReturn:
     except RuntimeError as error:
         logger.critical(f"ATTENTION::Received a RuntimeError while self terminating.\n{error}")
     sys.stdout.write(f"\rMemory consumed: {support.size_converter(0)}"
-                     f"\nTotal runtime: {util.time_converter(second=time.time() - shared.start_time)}")
+                     f"\nTotal runtime: {support.time_converter(second=time.time() - shared.start_time)}")
 
 
 def sleep_control() -> bool: