+++ modules/meetings/events.py
@@ -14,7 +14,7 @@ from modules.database import database
 from modules.logger.custom_logger import logger
 from modules.models import models
 from modules.retry import retry
-from modules.utils import shared, support
+from modules.utils import shared, util
 
 db = database.Database(database=models.fileio.base_db)
 
@@ -91,7 +91,7 @@ def events_gatherer() -> str:
     local_events = local_events.replace(", date ", " rEpLaCInG ")
     event_time = local_events.split("rEpLaCInG")[1:]
     event_name = local_events.split("rEpLaCInG")[0].split(", ")
-    event_name = support.remove_duplicates(input_=event_name)
+    event_name = util.remove_duplicates(input_=event_name)
     count = len(event_time)
     [event_name.remove(e) for e in event_name if len(e) <= 5] if count != len(event_name) else None
     event_status = f"You have {count} events in the next 12 hours {models.env.title}! " if count > 1 else ""